terms-of-service-alerts.js -> terms-of-service-alerts.ts
